Added inverse position (ik6_gen) and velocity (ivk6_gen) kinematics.
These are files as generated by matlabFunction(). The inverse velocity
kinematics took >1hr to write to a file, so don't try to do it again.
The two functions take in transformation matrices (ivk6 also takes
the time derivative of the transformation matrix). The returned
matrix will be 6x2, where each row corresponds to a different joint
and the two columns to the two valid joint configurations.

These generated files are much faster than using the symbolic subs()
function.